Migrate the tricky sso callback endpoint, remove old approach
