+++ server/service/transport_sessions.go
@@ -1,21 +1,1 @@
 package service
-
-import (
-	"context"
-	"net/http"
-
-	"github.com/fleetdm/fleet/v4/server/contexts/ctxerr"
-	"github.com/fleetdm/fleet/v4/server/sso"
-)
-
-func decodeCallbackSSORequest(ctx context.Context, r *http.Request) (interface{}, error) {
-	err := r.ParseForm()
-	if err != nil {
-		return nil, ctxerr.Wrap(ctx, err, "decode sso callback")
-	}
-	authResponse, err := sso.DecodeAuthResponse(r.FormValue("SAMLResponse"))
-	if err != nil {
-		return nil, ctxerr.Wrap(ctx, err, "decoding sso callback")
-	}
-	return authResponse, nil
-}
